requirements.txt
selenium>=4.0.0
python-dotenv>=1.0.0
keyboard>=0.13.5